feat: add version argument
